Refactor the code from last session to add a post controller
